+++ Sources/PippinLibrary/Foundation/NSOperation/CompoundOperation.swift
@@ -10,7 +10,7 @@ import Foundation
 enum CompoundOperationError: Error {
     case suboperationFailure(String, [Error])
     
-    public static let nsErrorDomain = String(asRDNSForCurrentAppWithSubpaths: ["error.domain.compound-operation"])
+    public static let nsErrorDomain = String(asRDNSForPippinSubpaths: ["error.domain.compound-operation"])
     
     func nsError() -> NSError {
         switch self {
@@ -31,14 +31,14 @@ enum CompoundOperationErrorUserInfoKey: CustomStringConvertible {
         switch self {
         case .underlyingErrors: name = "underlying-errors"
         }
-        return String(asRDNSForCurrentAppWithSubpaths: ["error.user-info-key.compound-operation", name])
+        return String(asRDNSForPippinSubpaths: ["error.user-info-key.compound-operation", name])
     }
 }
 
 open class CompoundOperation: AsyncOperation {
     @objc public static let CompoundOperationErrorCodeSuboperationFailed = 1
 
-    static let CompoundOperationCountingQueueLabel = String(asRDNSForCurrentAppWithSubpaths: ["dispatch-queue", "label", "compound-counting-queue"])
+    static let CompoundOperationCountingQueueLabel = String(asRDNSForPippinSubpaths: ["dispatch-queue", "label", "compound-counting-queue"])
     
     @objc public let compoundQueue = OperationQueue()
     var completedOperations = 0