+++ Sources/PippinLibrary/Foundation/String/String+rDNS.swift
@@ -8,6 +8,16 @@
 
 import Foundation
 
+private class PippinDummyRDNSBundleSearchClass {}
+
+extension String {
+    init(asRDNSForPippinSubpaths subpaths: [String]) {
+        precondition(subpaths.count > 0)
+        precondition(subpaths.filter({ $0.count > 0 }).count > 0)
+        self = String(asRDNSWithSubpaths: [Bundle(for: PippinDummyRDNSBundleSearchClass.self).identifier] + subpaths)
+    }
+}
+
 public extension String {
     init(asRDNSWithSubpaths subpaths: [String]) {
         precondition(subpaths.count > 0)