+++ Sources/PippinCore/CHANGELOG.md
@@ -6,6 +6,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/) 
 
 # [Unreleased]
 
+### Fixed
+
+- Use correct bundle ID in `DefaultDefaults`, `LogLevel`, `LaunchArgument` and `EnvironmentVariable` descriptions, as well as errors and internal components.
+
 ## [2.0.1] 2020-04-05
 
 ### Fixed